+++ crates/compiler/src/driver_constants.rs
@@ -174,7 +174,7 @@ pub const MAX_SOUND_EFFECTS: usize = 254;
 
 pub const MAX_N_PITCHES: usize = 256;
 
-pub const COMMON_DATA_POINTERS_SIZE: usize = 12 * 2;
+pub const COMMON_DATA_POINTERS_SIZE: usize = 11 * 2;
 pub const COMMON_DATA_HEADER_SIZE: usize = COMMON_DATA_POINTERS_SIZE + 4;
 
 pub const COMMON_DATA_PITCH_TABLE_OFFSET: usize = 20;
@@ -182,7 +182,7 @@ pub const COMMON_DATA_DIR_TABLE_OFFSET: usize = COMMON_DATA_HEADER_SIZE;
 
 pub const COMMON_DATA_BYTES_PER_DIR: usize = 4;
 pub const COMMON_DATA_BYTES_PER_PITCH: usize = 2;
-pub const COMMON_DATA_BYTES_PER_INSTRUMENT: usize = 4;
+pub const COMMON_DATA_BYTES_PER_INSTRUMENT: usize = 3;
 pub const COMMON_DATA_BYTES_PER_SOUND_EFFECT: usize = 4;
 pub const COMMON_DATA_BYTES_PER_SFX_SUBROUTINE: usize = 2;