+++ crates/compiler/src/errors.rs
@@ -479,6 +479,7 @@ pub struct SampleAndInstrumentDataError {
 
 #[derive(Debug)]
 pub enum CommonAudioDataError {
+    // ::TODO remove::
     TooManyInstrumentsAndSamples(usize),
     TooManyBrrSamples(usize),
     TooManySfxSubroutines(usize),